Fix highlight xss bug
